# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Store Service API
  version: 1.0.0

paths:
  /products:
    get:
      summary: Get a list of products
      description: Retrieves a paginated list of products.
      operationId: getProducts
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: A paginated list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '204':
          description: No products found
        '400':
          description: Invalid request
        '500':
          description: Internal server error
    post:
      summary: Create a new product
      description: Creates a new product with price in EUR. Price in USD will be computed automatically.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateDto'
      responses:
        '201':
          description: Product created
          headers:
            Location:
              description: URI of the newly created product
              schema:
                type: string
        '400':
          description: Invalid request
        '500':
          description: Internal server error

  /products/{code}:
    get:
      summary: Get product by code
      description: Returns the details of a product
      operationId: getProduct
      parameters:
        - in: path
          name: code
          required: true
          schema:
            $ref: '#/components/schemas/Code'
          description: code of the product to retrieve
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponseDto'
        '404':
          description: Product not found
        '500':
          description: Internal server error

components:
  schemas:
    ProductCreateDto:
      type: object
      required:
        - code
        - name
        - priceEur
      properties:
        code:
          $ref: '#/components/schemas/Code'
        name:
          type: string
          example: "Wireless Mouse"
          description: Name of the product
        priceEur:
          type: number
          format: decimal
          minimum: 0
          example: 29.99
          description: Price in EUR
        isAvailable:
          type: boolean
          default: true
          example: true
          description: Availability of the product

    ProductResponseDto:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/Code'
        name:
          type: string
          example: "Wireless Mouse"
        priceEur:
          type: number
          format: decimal
          example: 29.99
        priceUsd:
          type: number
          format: decimal
          example: 32.50
        isAvailable:
          type: boolean
          example: true
    
    ProductPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponseDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    Pageable:
      type: object
      properties:
        page:
          type: integer
          description: Page number (0-based)
        size:
          type: integer
          description: Page size
        sort:
          type: string
          description: Sorting criteria in the format (asc|desc).
          
    Code:
      type: string
      minLength: 10
      maxLength: 10
      example: "ABC1234567"
      description: Unique product code